Trapped 'Ctrl + C' to protect 'wait' in recording Australian National Radio news.
When listening to funkRadio, it is convenient to move to the next song or news broadcast by typing Ctrl + C. However, that shortcut interferes with the 'wait' command in the abcradnatnews function. In order to enable recording of ABC National Radio news, the shortcut was trapped in the abcradnatnews function. But then 'wait' needed to be replaced by the line 'until wait;do :;done'. Thanks to superusers at  
https://superuser.com/questions/1719758/bash-script-to-catch-ctrlc-at-higher-level-without-interrupting-the-foreground